redesign ui top done
